Rename module with Suspend constants
